+++ src/test/java/org/openrewrite/cucumber/jvm/RegexToCucumberExpressionTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cucumber.migration;
+package org.openrewrite.cucumber.jvm;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
